@@ -25,7 +25,7 @@
 		"autoprefixer": "^10.3.3",
 		"postcss": "^8.3.6",
 		"tailwindcss": "^2.2.7",
-		"typescript": "^4.3.5",
+		"typescript": "^4.4.2",
 		"vite": "^2.5.1",
 		"vite-plugin-fonts": "^0.2.2",
 		"vue-tsc": "^0.3.0"
